We now get the fields of a table
If the fields are not provided we query the db to get the list of column names. This is need to insure that when we are getting the geom
field that it is gotten in wkb format (via the ST_AsBinary function).

Also, when getting the id field, we make sure to convert it to an int64. We no longer assume it will be in int64; but only that it will be
an int of some sort.