Added basic support for 4326 Datasource.
Add functions to convert a geometry in 4326 to WebMercator and back. Added support to the postgis provider to do the conversion.
I don't think this is the right place for the conversion to take place, but is good enought for now. It should live in the feature
but that requires the features to know the encoding of the geometries, instead of assuming WebMercator.

I'm using the more correct rather then the faster conversion algo.

Relates to issue #47