Initial Simplification code.
This code works and simplify the points as one zooms out. The performance of the code is slowed down as
it is having to loop through all the points one more time. May be able to optimize this later. Also, there is a possible bug with features that become too small to show. We return nil, but not sure how the encoding layer is going to deal with this. It should ignore this, but this could mean that a feature totally disappears. In which case the layer it's self should not be rendered. Still have to test this.

Also, missing are test cases.