Clipping of a line string is now working.
Got all tests to pass. Need to create more tests, especially for edge cases.
Happy with the overall structure of the code. Code is better organized and
easier to understand what is going on.

There are two hacks one should be aware off.

1. Using a hash pointer values to check if we have already seen a point. If we
   have, we bail the walk because we are stuck in a cycle.
2. Check to see if we saw the same point value as the starting inbound point.

The other hack is, that if the number of values is less than six, we discard it
as a valid segment  - these are possible because of the hacks mentioned above.