Mostly Prep work for Validate Polygon code
This is mostly prep work for the IsValidate and MakeValidate functions we need
for polygons. I think this is one of the issues with the polygons, that the
multipolygons we get are not always OGC complient and we are making the assumption
that is the case. It has also been pointed out that after scaling it is possible
for certain small polygons to flip directions, so we need to insure this does
not happen.