Plumbed the context struct through the tile generation. [issue-111]
Need to plumb the context for cancellation throught tile generation.
I did not plumb it into the scaling, or clipping functions as we can
cancel between the operations, that way we don't incure extra check
when comparing lot's of points.

If we need to plumb it futher down later, we will be able to.

Fixes #111