Add flags and build tags for profiling. [issue-122]
When we are not profiling we really don't want to compile in
all the profiling code. So, we are using build flags to conditionally
compile in profiling code as needed.

When compiled with profiling, a new option is enabled: profile.mode

This can take four values: cpu, mem, mutex, block
These will allow those types of profiling profile to be created.

To build with profiling use the -tag="profile" flag.