[issue-56] Updated system to use simplification.
Simplification will not be applied to zoom levels 14 and above. This
mirrors https://github.com/mapbox/geojson-vt. Layers can now opt out
of simplification or set their own max zoom level at which point
simplification will not be applied. (Default is 14). This feature
is not currently exposed in the configuration.