[issue-107] Expanded bbox for grabbing features.
This fix does not make the size of the tile buffer available in the
config file.

We believe the reason for some of the jagged lines that occurs has to
do with features that are missing from the buffer region of the tile.
When grabbing features to display, we were only grabbing feature that
intersected the bounding box of the tile, instead of the bounding box
of the tile plus a buffer zone. We query for all the features we need,
including the ones that only intersect with the buffer
zone.

Some of the issues reported in #107, we believe is caused by these
miscalculated bounding boxes.

Additional clean up is required to remove unused code.