[issue-107] Got all tests to pass.
A lot of tests were wrong due to the misunderstanding by me of how the
bounding box, tile extent (know as a layer extent in the mvt spec) and
tiles buffers worked. Fixed all of this. Fixed the test cases, however
some of the tests aren't the best till we get #224 done. This is because
of doing division on float64.

Going to Pixels there is going to be trancations, and then going back
will cause a bit of a loss. This is fine for calculating the buffer
areas, but causes issues with using that technic for encoding.