conditional cgo compilation #161
If one now uses `CGO_ENABLED=0` the system will now disable gpkg
support. The provider will not register it's self with the system
and a dummy package that returns and err if the NewProvider method
is called will be returned instead. Functions that don't require cgo
are still available to other programs. However, note that functions
that require cgo (any function interacting with sqlite) will not be
available.