Refactored Atlas to enable testing.
Removed, reliance on Atlas global in the server package. Refactored
Atlas, so that an Atlas can be passed to both HandleMapZXY and
HandleMapLayerZXY, this will allow a single test to be able to test
with different versions of Atlas, without the need to override the
default Atlas, in the the Atlas package.

Made Default Atlas private, forcing users of it to to go through
the package api.