gzip compression support #438
implements gzip compression on successful responses
when the Accept-Encoding header is properly set.

When leveraging a cache back end for caching tiles
the tile data will be compressed prior to writing
to the cache.

If an incoming request does not support gzip compression
the response data will be decompressed prior to responding.

NOTE: this commit will break all current caches! It's critical
that all caches be purged and rebuilt!