[#529] Changed the way cache command works.
issue #529 discussion outline a better way for the cache to
handle explitic and non-explicit tile lists. This commit
breaks up the cache command into seperate commands and
allows one to specify a tilelist, if the min or max zoom
is not given for a tilelist, then that tile list is treated
as an explicit list. However, if they are provided, then the
system will find all tiles within the given zoom range that
would be effected by the tiles in the tile list and reseed or
purge them accordingly.

This builds off of #533
closes #533