Changes to explicit tiles list as per #529 (#547)
* Adds the explicit-list option for seeding
* Added kill after 60 seconds.
* vendor v3.2.0 of pgx
* Give units when reporting large tiles. closes #494 (#542)
* Changed the way cache command works.
issue #529 discussion outline a better way for the cache to
handle explicit and non-explicit tile lists. This commit
breaks up the cache command into separate commands and
allows one to specify a tile list, if the min or max zoom
is not given for a tile list, then that tile list is treated
as an explicit list. However, if they are provided, then the
system will find all tiles within the given zoom range that
would be effected by the tiles in the tile list and reseed or
purge them accordingly.
closes #533
* Updated provider clean up messages.
On cleanup the provider cleanup routine would report it was
cleaning up the provider even if there weren't any providers
to clean up. Now do a quick check before printing out the
message.
* added environmental variable ( TEGOLA_PPROF_MUTEX_RATE ) to enable
mutex profiling. Value must be between 1 to 10000.
* added environmental variable ( TEGOLA_PPROF_BLOCK_RATE ) to enable
block rate profiling.  Value must be between 1 and 10000.
* Replaced strings.Builder (no supported in go1.9.x) with bytes.Builder
`strings.Builder` was introduced in go 1.10.x, so we cannot use it till
we drop support for go 1.9.x.
* fixed incorrect comment in pprof file