V0.8.x (#548)
* fix for PGX numeric types sharing pointer reference (#487)
* support for lambda during CI build
* added TEGOLA_ prefix to SQL_DEBUG env var. closes #489 (#491)
* provider/gpkg: various fixes #486
- provider/gpkg: fix index query for geomFieldname != geom
- rtree table name is a combination of the table name and the geometry
column name.
- provider/gpkg: use aliases and quotes in query for all column names
- fixes 'ambiguous column name' errors when column names in data table
colide with columns in rtree index table (id, minx, etc).
- provider/gpkg: improve column names extraction
- Support column names without quotes.
* provider/postgis: replace Mapnik pixel_width/height and scale_denominator SQL tokens
* added uppercaseTokens func to support case insensitive tokens within ! bangs. mapnik uses lowercase tokens (i.e. !pixel_width!) and tegola uses upper case tokens (i.e. !PIXEL_WIDTH!). this allows for either to work.
* Issue 459 mime type (#504)
* cache/s3: added default tile mime-type of application/vnd.mapbox-vector-tile
* additional test case for provider/gpkg -> extractColsAndPKFromSQL
* documenting new SQL tokens for postgis
* provider/postgis: support for SQL sub-queries #467
* provider/postgis: fix mix-up in SQL token documentation (#510)
* provider/postgis - close connection pool (#515)
* added Cleanup routine to postgis to explicitly close the connection pool on shutdown.
* better error management for http server.ListenAndServe
* Issue 438: gzip compression (#509)
implements gzip compression on successful responses
when the Accept-Encoding header is properly set.

When leveraging a cache back end for caching tiles
the tile data will be compressed prior to writing
to the cache.

If an incoming request does not support gzip compression
the response data will be decompressed prior to responding.

NOTE: this commit will break all current caches! It's critical
that all caches be purged and rebuilt!
* added Content-Length header support for gzipped tiles
* remove Content-Length header on gzip decompress
* moved MVT MimeType to a const in the mvt package
* updated lambda instructions
* added postgis log for 0 len geom bytes
* removed the unique ID check to match the mvt spec which says IDs SHOULD be unique but is not a MUST. closes #338 (#522)
* reworked mvt tests to closer align to the new testing structure and remove the dependency on tbltest package.
* removed default id (gid) for postgis provider.
* fixed mvt layer.AddFeature append.
* null id_fieldname handling in postgis provider.
* fixes std stream targets... msgs intended for stdout (severity level WARN and below) now actually get sent to stdout instead of everything being sent to stderr
* added debug log statements to show sql and size of record sets in gpkg.TileFeatures() set log severity to DEBUG in gpkg pkg
* Issue 519: Support for configurable HTTP response headers (#537)
* config: webserver section: remove CORSAllowedOrigin, add Headers subsections
* Add blacklisted headers
* Add ability to redefine CORS headers from [webserver.headers] config
* Add HeadersHandler middleware, enable it to tegola and tegola_lambda
* Add [webserver.headers] section in README.md
* provider/postgis: fix uppercaseToken for queries with != etc. (#528)
* better error checking for provider/postgis. closes #538
* updating docker images. closes #385
We had a redundant Dockerfile which would have to maintained as well
as the Dockerfile template that is used by our CI system. Removed
Dockerfile, extracted out the Dockerfile generation and added a README
to enable one to just generate the Dockerfile.
* Add read only runtimeParm to Postgress (#541)
* Add default_transaction_read_only to provider/postgis runtime params
* added application_name to provider/postgis runtime params
* Improve display of tile rendering times. closes #484 (#543)
* Give units when reporting large tiles. closes #494 (#542)
* Changes to explicit tiles list as per #529 (#547)
* Adds the explicit-list option for seeding
* Added kill after 60 seconds for cache seed / purge.
* vendor v3.2.0 of pgx
* Changed the way cache command works.
issue #529 discussion outline a better way for the cache to
handle explicit and non-explicit tile lists. This commit
breaks up the cache command into separate commands and
allows one to specify a tile list, if the min or max zoom
is not given for a tile list, then that tile list is treated
as an explicit list. However, if they are provided, then the
system will find all tiles within the given zoom range that
would be effected by the tiles in the tile list and reseed or
purge them accordingly. closes #533
* Updated provider clean up messages.
On cleanup the provider cleanup routine would report it was
cleaning up the provider even if there weren't any providers
to clean up. Now do a quick check before printing out the
message.
* added environmental variable ( TEGOLA_PPROF_MUTEX_RATE ) to enable
mutex profiling. Value must be between 1 to 10000.
* added environmental variable ( TEGOLA_PPROF_BLOCK_RATE ) to enable
block rate profiling.  Value must be between 1 and 10000.
* fixed incorrect comment in pprof file