Use a hard-coded zoom level when inspecting geom type
This resolves #343, although not in the most ideal way. I've been
running with this patch on openinframap.org for a few months so I
think it's worth documenting it.

As far as I'm aware, the only situation in which this would cause a
problem is if a layer's SQL has a zoom filter which filters out all
objects at zoom 24, but displays them at lower zooms.

I'm not sure if this is a common situation (it isn't in my
configuration), and I don't necessarily expect this patch to be
merged. I'm providing it so it's available to others as an option, in
the absence of a better solution.