Move geometry processing out of mvt package
The mvt encoding package contained core geo processing routines
(i.e. simplification, make valid, scaling). These routines have
been carved out of the mvt pacakge and moved into the atlas
map Encode() step. This allows the mvt package to focus entirely
on encoding.

This is a transitional step which positions tegola to adopt the
new make valid routine which is currently under development in the
geom package. This transition also positions tegola to adopt the
new simplification routine which is in the geom package. The mvt
package will soon be carved out of tegola and also moved to the
geom package.

During the refactor, a fair amount of code was purged from the mvt
package and many function signatures were able to reduce the number
of parameters they required. Additionally the mvt feature
encoding tests were able to fixed to match the specifications
supplied tests and previously breaking tests are now passing.

closes #224