Added configurable webserver URI Prefix to
Added the webserver config property uri_prefix which allows for
setting an API route prefix. This is commonly used when tegola
is run behind a reverse proxy (i.e. example.com/tegola). Certain
routes (i.e. capabilities) build URLs which are included in the
response and then passed back to the proxy. Without the ability
to set a URI prefix, the URLs in the capabilities responses
will not be useable by the consuming client.

Setting the webserver hostname now ignores the server port. It's
assumed that if the user is setting the hostname they want full
control of it.

closes #136