These are the change I propose.
The big thing is let's make the go-generate version the same as running it by hand or through the ci.

To do this, I'm proposing a small go build file (ui/build.go) which will run the necessary command to build out the ui and update the bindata.go file.

Next make vierwer_gen.go make use of this build.go file.