Separate Linux and MacOS CI release build jobs
Compiling MacOS release builds on Linux, as the GitHub Actions CI job
does now, means cgo isn't enabled, which means the go-sqlite3 package
isn't compiled, which means no GeoPackage support in the MacOS builds.

This commit separates the jobs so that Linux builds are compiled on the
ubuntu-latest runner while MacOS builds are compiled on the macos-latest
runner. This means no cross-compilation is required and cgo can run
happily. For further details, see the discussion at
https://github.com/go-spatial/tegola/issues/736.