Make sure to strip "/" for static files if URIPrefix ends with a "/"
When serving up static files we need to strip any proxy prefixes we may add onto the URL.
We forgot to check to see if the prefix had a "/" at the end of it, and if it did made sure
to strip it out, as static files should always start with a "/". However, if we did not remove
the "/" at the end of the prefix that we then use to trim the url; we would end up trimming off
the starting "/". This fixes that.

Fixes #735