Refactor providers/postgis to use pgx v4
* refactored providers/postgis to use the pgx4 client.
Support for Postgres versions > 12 is now possible.

* provider/postgis: Properly wrap errors in messages by
moving from using %v -> %w when returning errors
in messages.

* Added error string check for context.Canceled.
The underlying net.Dial function is not properly
reporting context.Cancel errors. Becuase of this,
a string check on the error is performed. There's
an open issue for this and it appears it will be
fixed eventually but for now we have this check
to avoid unnecessary logs.

Related issue: https://github.com/golang/go/issues/36208

* added ctxErr() check thewill check if the supplied
context has an error (i.e. context canceled) and if so,
return that error, else return the supplied error.
This is useful as not all of Go's stdlib has adopted
error wrapping so context.Canceled errors are not always
easy to capture.

closes #748