Automate the generation of Tags slice
Before when adding a new build tags target we need to make sure
that the new tag was added to the build.Tags slice, so that at
runtime we knew what tags were used (and thus features) to build the
software.

This can be error prone, as one can forget to add the tag and it's
negative version to the tags list.

By doing a static analyze of the code we can determine the tags that
are possible and generate the necessary files.