build.Import is not finding package for some reason
Moved away from build.Import as it was not finding the
package when given the sourceDirectory directly. Something
has changed in the API, not sure what. Using the `go list`
command instead.

This is fine as we where depending on it already to get the list
of all the packages in a repository. So, not we use it to get
the source directory for a package and the packages Name.

This is then used with `build.ImportDir` to get the full information
about the package.