cache: allow to set a local proxy endpoint for s3
If Tegola is configured to use a s3 cache endpoint like
http://localhost:1234, the AWS request signing step
is not done corretly and the real S3 endpoint will
return a HTTP 403 stating that the request signing
header value doesn't match what it expects.

This is a typical use case in k8s, where the HTTP
connection go through a local sidecar/proxy. In our case,
we have an internal S3 endpoint (implemented via Openstack
Swift) and we noticed an increase in CPU usage in the pod
running tegola when the TLS cert of the S3 endpoint changed
(probably due to high requirements for the Cipher suite).
We would like to implement in the local proxy connection pooling
and other similar performance improvements, without
necessarily changing any of the Tegola's code or settings.