When generating the seed list, we need the grid for the bounds
Added the ability to specify the srid to use. It defaults to 4326.

Our old system normalized the bounds to grid, this meant that people could provide a bounds without the associated srid as
long as it was either 4326 or 3853. With the change over, we
were defaulting to 3853. This was the wrong default. Chanced the default 4326, and added ability to specify the srid.

We will check to see if the SRID is supported by proj. For now the only srid's that are supported are 4326 and 3853.