ui: Fix inspector CSS. Fix custom map control
After the viewer was migrated from mapboxgl to maplibregl
the associated CSS references were not updated. This caused
the styling for the inspector to use default browser styles.
The style references have now been corrected.

Additionally the debug tile line map control implementation was
was using mapboxgl CSS classes so the control was not being added
to the map. This has been fix.